Edited DB structure and value types

+++ models/vote.go
@@ -3,5 +3,5 @@ package models
 //Vote - struct of vote
 type Vote struct {
 	Nickname string `json:"nickname"`
-	Voice    int    `json:"voice"`
+	Voice    int16  `json:"voice"`
 }
